Add hacky support for clickable a tags in SVGs:
This copies a lot of the code from the anchor tag element, but modifies it to
work for SVGs. I couldn't directly implement the `follow_the_hyperlink` method
because of how inheritence conflicts (they both specify `href`, which conflict)
though I should be able to put this into a util wrapper and call it that way.